fix: move ttf calculator into ttf services directory
